refactor(WebAPI): restructure and split fake data for download tasks
