+++ src/Domain/Entities/Plex/PlexAccount.cs
@@ -85,17 +85,17 @@ public class PlexAccount : BaseEntity
 
     /// <summary>
     /// The user has the option to provide their own token to authenticate with plex.tv.
-    /// This is not the same as the auto filled AuthenticationToken when provided by the username and password
+    /// This is not the same as the auto-filled AuthenticationToken when provided by the username and password
     /// </summary>
     [Column(Order = 13)]
-    public required string CustomAuthenticationToken { get; init; }
+    public required string CustomAuthenticationToken { get; init; } = string.Empty;
 
     /// <summary>
     /// The general plex authentication token used to retrieve account data such as the <see cref="PlexServer" />s the
     /// account has access to. This is filled in by the PlexAPI
     /// </summary>
     [Column(Order = 14)]
-    public required string AuthenticationToken { get; set; }
+    public required string AuthenticationToken { get; set; } = string.Empty;
 
     /// <summary>
     /// If this is a main account then it will get a lower priority when downloading media which a non-main account also has access to.
